+++ fishbowl_common/__init__.py
@@ -1,7 +1,15 @@
 from fishbowl_common.ArgumentProvider import ArgumentProvider
 from fishbowl_common.PatchNotes import PatchNotes
 from fishbowl_common.SettingsRepository import SettingsRepository
-from fishbowl_common.UpdateChecker import ReleaseAsset, UpdateChecker, UpdateCheckResult
+from fishbowl_common.UpdateChecker import (
+    CHECK_ERROR_HTTP,
+    CHECK_ERROR_NETWORK,
+    CHECK_ERROR_RATE_LIMITED,
+    CHECK_ERROR_RESPONSE,
+    ReleaseAsset,
+    UpdateChecker,
+    UpdateCheckResult,
+)
 from fishbowl_common.UpdateCoordinator import UpdateCoordinator, UpdateDisplay
 from fishbowl_common.UpdateDownloader import UpdateDownloader
 from fishbowl_common.UpdateInstaller import UpdateInstaller
@@ -9,6 +17,10 @@ from fishbowl_common.version_utils import compare_versions, parse_version
 
 __all__ = [
     "ArgumentProvider",
+    "CHECK_ERROR_HTTP",
+    "CHECK_ERROR_NETWORK",
+    "CHECK_ERROR_RATE_LIMITED",
+    "CHECK_ERROR_RESPONSE",
     "PatchNotes",
     "ReleaseAsset",
     "SettingsRepository",